+++ prototype2/gdgem/nmx/Eventlet.h
@@ -20,7 +20,7 @@ public:
   static constexpr adc_type adc_max_val{std::numeric_limits<adc_type>::max()};
 
 public:
-  uint64_t time{0};
+  double time{0};
   uint8_t plane_id{0};
   strip_type strip{0};
   adc_type adc{0};
